backend/requirements.txt
fastapi
uvicorn
python-multipart
pdf2image
google-auth
google-auth-oauthlib
google-auth-httplib2
pyOpenSSL
python-dotenv
Pillow
google-cloud-storage
google-cloud-firestore
bcrypt>=3.2.0
aiofiles>=23.2.1

pytest>=8.0.0
requests>=2.31.0
